// frontend/src/types/common/animationTypes.ts
// Define possible general animations
export type AnimationType =
  | "slide-up"
  | "slide-down"
  | "slide-left"
  | "slide-right"
  | "scale-up"
  | "scale-down";

// Define possible animation effects
export type AnimationEffectType =
  | "shake"
  | "bounce"
  | "pulse"
  | "heartbeat"
  | "rubberBand"
  | "wobble"
  | "jello"
  | "flash"
  | "bounceIn"
  | "bounceOut";